0.9.6

+++ src/UfpOptimizer.js
@@ -4,25 +4,21 @@ const optimHTML = require('./HtmlOptim')
 const optimizeCSS = require('./CssOptim')
 const optimZIP = require('./ZipOptim')
 const fs = require('fs-extra')
+const helper = require('./Helper')
 
 var UfpOptimizer = {}
 
 UfpOptimizer.execute = function (settings) {
-    console.log('* step0 - copy: started')
-    UfpOptimizer.copy(settings)
-    console.log('* step0 - copy: finished')
+
+
     // app.optimizeCSS()
-
-    console.log('** step1 - image/html/css: started')
-    Promise.all([
-        UfpOptimizer.optimizeImages(settings),
-        UfpOptimizer.optimizeHTML(settings)]).then(function () {
-        console.log('** step1 - image/html/css: finished')
-
-        console.log('*** step2 - compression: started')
-        var result = UfpOptimizer.zip(settings)
-        console.log('*** step2 - compression: finished')
-        return result
+    return UfpOptimizer.copy(settings).then(function (result) {
+        return Promise.all([
+            UfpOptimizer.optimizeImages(settings),
+            UfpOptimizer.optimizeHTML(settings)]).then(function () {
+            var result = UfpOptimizer.zip(settings)
+            return result
+        })
     })
 }
 
@@ -31,25 +27,57 @@ UfpOptimizer.getDefaultSettings = function () {
 }
 
 UfpOptimizer.copy = function (settings) {
-    // prepare
-    fs.removeSync(settings.outputDir)
-    fs.mkdirSync(settings.outputDir)
-    fs.copySync(settings.inputDir, settings.outputDir)
+    return new Promise(function (fulfill, reject) {
+        console.log('* ufp-optimizer copy: started', settings.inputDir, '=>', settings.outputDir)
+
+        if (!fs.existsSync(settings.inputDir)) {
+            console.log('ERROR: input dir does not exist', settings.inputDir)
+            reject(settings.inputDir);
+            return;
+        }
+
+        if (settings.outputDir !== settings.inputDir) {
+            // prepare
+
+
+            fs.removeSync(settings.outputDir)
+            fs.mkdirSync(settings.outputDir)
+            fs.copySync(settings.inputDir, settings.outputDir)
+
+            console.log('* ufp-optimizer - copy: finished')
+
+        } else {
+            console.log('* ufp-optimizer - copy: finished did nothing')
+        }
+
+        fulfill()
+    });
+
 }
 
 UfpOptimizer.optimizeImages = function (settings) {
-    // optimize
+
+    console.log('** ufp-optimizer  - image/html/css: started')
     var files = fs.walkSync(settings.outputDir)
-    return optimImages.optimizeFileList(files, settings)
+    return optimImages.optimizeFileList(files, settings).then(function (result) {
+        console.log('** ufp-optimizer  - image/html/css: finished')
+        return result
+    })
+
 }
 
 UfpOptimizer.optimizeHTML = function (settings) {
+    console.log('** ufp-optimizer  - image/html/css: started')
     // optimize
     var files = fs.walkSync(settings.outputDir)
-    return optimHTML.optimizeFileList(files, settings)
+    return optimHTML.optimizeFileList(files, settings).then(function (result) {
+        console.log('** ufp-optimizer  - image/html/css: finished')
+        return result
+    })
 }
 
 UfpOptimizer.optimizeCSS = function (settings) {
+    console.log('** ufp-optimizer  - image/html/css: started')
     // optimize
     var files = fs.walkSync(settings.outputDir)
     return optimizeCSS.optimizeFileList(files.filter(function (entry) {
@@ -58,12 +86,19 @@ UfpOptimizer.optimizeCSS = function (settings) {
         htmlFiles: files.filter(function (entry) {
             return ['.htm', '.html'].indexOf(path.extname(entry)) > -1
         })
-    }, settings)
+    }, settings).then(function (result) {
+        console.log('** ufp-optimizer  - image/html/css: finished')
+        return result
+    })
 }
 
 UfpOptimizer.zip = function (settings) {
+    console.log('*** ufp-optimizer  - zip: started')
     var files = fs.walkSync(settings.outputDir)
-    return optimZIP.optimizeFileList(files, settings)
+    return optimZIP.optimizeFileList(files, settings).then(function (result) {
+        console.log('*** ufp-optimizer  - zip: finished')
+        return result
+    })
 }
 
 module.exports = UfpOptimizer
